unsigned long long frequency

+++ Huffman/ByteFrequency.cpp
@@ -3,7 +3,7 @@
 #include "ByteFrequency.h"
 #include "Types.h"
 
-ByteFrequency::ByteFrequency(unsigned int f, byte b) : frequency(f), byte_code(b) {
+ByteFrequency::ByteFrequency(unsigned long long f, byte b) : frequency(f), byte_code(b) {
     if (f < 0) {
         cerr << "[ByteFrequency]: Invalid frequency";
         exit(-1);
@@ -12,13 +12,13 @@ ByteFrequency::ByteFrequency(unsigned int f, byte b) : frequency(f), byte_code(b
 
 ByteFrequency::ByteFrequency() : frequency(0), byte_code(0) {}
 
-unsigned int ByteFrequency::get_frequency() const {
+unsigned long long ByteFrequency::get_frequency() const {
     return frequency;
 }
 
 byte ByteFrequency::get_byte_code() const { return byte_code; }
 
-void ByteFrequency::set_frequency(unsigned int f) {
+void ByteFrequency::set_frequency(unsigned long long f) {
     if (f < 0) {
         cerr << "[ByteFrequency]: Invalid frequency";
         exit(-1);
@@ -34,7 +34,7 @@ bool ByteFrequency::is_empty() { return frequency == 0; }
 ofstream& operator<<(ofstream& os, const ByteFrequency& bf)
 {
     os << bf.get_byte_code();
-    unsigned int f = bf.get_frequency();
+    unsigned long long f = bf.get_frequency();
     os.write(reinterpret_cast<const char*>(&f), sizeof(f));
     return os;
 }
