+++ Huffman/Decoder.cpp
@@ -31,7 +31,7 @@ void Decoder::decode(char* input_directory, char* output_directory)
         input.get(c);
 
         // frequency
-        unsigned int frequency_value;
+        unsigned long long frequency_value;
         input.read(reinterpret_cast<char*>(&frequency_value), sizeof(frequency_value));
         //input >> frequency_value;
